requirements.txt
tensorflow==2.15.0
pandas 
numpy 
scikit-learn
tensorboard
matplotlib
streamlit
scikeras
nltk
torch
torchvision
yfinance
gym
stable_baselines3[extra]
gymnasium
pygame
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
PyPDF2
transformers
sentence-transformers
faiss-cpu
